pyproject.toml
[tool.poetry]
name = "suppliers-sku-map"
version = "0.1.0"
description = ""
authors = ["Izbushko <izbaizbushka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
